Preserve state of spring-boot apps
